Create RootToNodeInBT.java
Base Case: Checking if root==null;

Check if root.data == x. If it is then make an arraylist and add root.data to it and return arrayList

else 
call recursion on left side. If recursion on left side doesnt return null then it means the desired element is in the left subtree.
so add root.data to the returned arrayList and return it

else 
do the same process for the right side

else 
the element is neither in left subtree nor right subtree. So return null